feat: convert dep-graph to legacy tree
Exposes a way to convert dep-graph's to the legacy tree format.

"Package manager" and "package type" must be handled out of band by the
consumer, as "package manager" exists only on dep-graphs and "package
type" exists only on legacy trees.

Given legacy tree's have a very ill-defined structure, the conversion
can be lossy. This commit adds test fixtures that cover all of the data
that is preserved in conversion. If something is not in a fixture, then
it is not guaranteed to be preserved (e.g. package licence metadata).

It is expected that future releases of this library will formalise and
preserve more of the semi-structured data that exists on legacy trees, as
needed.