fix: avoid stack overflow with spread operator
A customer cannot scan their project with a supposedly very large dependency graph because as soon as it reaches our back-end, we get a stack overflow error. It was identified that the error is due to the combination of Array.push() and the spread operator.

For reference: https://github.com/nodejs/node/issues/16870

To reproduce, create an array of size 2^17 then do [].push(...myArray);.

This fix replaces all uses of push() and spread with a simple loop.