feat: Add an optional limit to pkgPathsToRoot
This patch adds an optional limit to `pkgPathsToRoot`, which allows us
to selectively avoid getting all paths to a given package

This is particularly useful if you only want to get a few paths for
display purposes, for example.

I've pulled out the benchmarks from scripts and added a new one here to
demonstrate the speed differences:

```
./benchmark/pkg-paths-to-root.ts

Running "pkgPathsToRoot" suite...
Progress: 100%

  with limit of 1:
    33 708 ops/s, ±1.53%   | fastest

  with limit of 100:
    916 ops/s, ±0.92%      | 97.28% slower

  with limit of 10,000:
    7 ops/s, ±6.18%        | slowest, 99.98% slower
```

This particular fixture has some 7 million paths, which on my laptop
takes up to ~6gb of memory and around 2 minutes. With a limit I can get
a single path in less than 1 millisecond