feat: ignore pruned nodes
In ecosystems where the dep-graph contains pruned nodes, the dep-graph
library functions like pkgPathsToRoot and countPathsToRoot can safely
ignore those pruned nodes.

This is because there should be a path to a non-pruned node, the
effective path.

Ignoring pruned paths has a number of advantages. Consuming applications
are presented with less paths to traverse when operating over paths.
When displaying dep-graph path information there is less noise as only
the effective paths are shown.